Trait samp_sdk::cell::repr::AmxCell

pub trait AmxCell<'amx> where
    Self: Sized, {
    fn as_cell(&self) -> i32;

    fn from_raw(_amx: &'amx Amx, _cell: i32) -> AmxResult<Self>
    where
        Self: 'amx,
    { ... }
}

AmxCell trait is a core trait of whole SDK. It shows that value can be borrowed (or copied if it's a primitive) from AMX and passed to it.

Required methods

fn as_cell(&self) -> i32

Get a value which can be passed to AMX.

Loading content...
Provided methods

fn from_raw(_amx: &'amx Amx, _cell: i32) -> AmxResult<Self> where
Self: 'amx,

Loading content...