Crate samp_sdk

SA:MP SDK

samp_sdk is a Rust lang bindings for original C SA:MP SDK

Re-exports

pub extern crate lazy_static;

Modules

amx

args

consts
AMX Constants.

data

macros
Some useful macros to easy access and define natives.

types

Macros

args_count
Get count of passed arguments in a native.

define_native
Define native and hide raw C export functions.

exec
Executes AMX::exec with given arguments.

exec_native
Finds a native function and executes AMX::exec with given arguments.

exec_public
Finds a public and executes AMX::exec with given arguments.

expand_args

get_array
Get a slice (an array) from arguments.

get_string
Gets a string from a raw pointer to Cell.

lazy_static

log
Useful macro to log to SA:MP server output.

natives
Clear macros that makes a new Vec<AMX_NATIVE_INFO>.

new_plugin
Hides ugly C code from your eyes.